# lib/api-zod/src/generated/types/blendInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infinite Us — couples relationship platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BlendInput {
  partner1: string;
  partner2: string;
  style?: string;
}