# lib/api-zod/src/generated/types/bond.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infinite Us — couples relationship platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Bond {
  id: string;
  name: string;
  code: string;
  createdBy?: string;
  memberCount?: number;
  createdAt: string;
}