# lib/api-zod/src/generated/types/journeyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infinite Us — couples relationship platform API
 * OpenAPI spec version: 0.1.0
 */

export interface JourneyUpdate {
  completedDays?: number;
  currentSeason?: string;
}