# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infinite Us — couples relationship platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomPhase } from './roomPhase';
import type { RoomStateJson } from './roomStateJson';

export interface Room {
  id: string;
  joinCode: string;
  coupleId: string;
  phase: RoomPhase;
  /** @nullable */
  currentCardId?: string | null;
  /** @nullable */
  stateJson?: RoomStateJson;
  createdAt: string;
  expiresAt: string;
}