# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infinite Us — couples relationship platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  coupleId: string;
  season?: string;
  cardsPlayed: number;
  cardsDrifted?: number;
  durationSeconds?: number;
  thisOrThatMatches?: number;
  thisOrThatTotal?: number;
  scoreEarned: number;
  /** @nullable */
  moodEmoji?: string | null;
  /** @nullable */
  momentTag?: string | null;
  cardTypes?: string[];
}